Create: 비회원 접근 제한 인터셉터

+++ goodee/src/main/java/com/goodee/proj/account/AccountService.java
@@ -69,8 +69,10 @@ public class AccountService {
 		
 		if (attach != null && !attach.isEmpty()) {
 			FileDTO oldFile = accountDAO.detailAttach(accountDTO.getAccountNumber());
-			fileService.deleteFile(oldFile);
-			result = accountDAO.deleteAttach(oldFile);
+			if (oldFile != null) {
+				fileService.deleteFile(oldFile);
+				result = accountDAO.deleteAttach(oldFile);
+			}
 			
 			String fileName = fileService.saveFile(FileService.ACCOUNT, attach);
